enclose tmux aliases in conditional that checks $TMUX

+++ .bashrc.d/tmux.sh
@@ -1,7 +1,11 @@
 # .tmux.sh
 # Shell commands for use with tmux
 
-# these are named to align with vim naming conventions
-alias vsplit="tmux split-window -h"
-alias hsplit="tmux split-window -v"
-alias close="tmux kill-pane -t"
+if [ -n "$TMUX" ]; then
+
+  # these are named to align with vim naming conventions
+  alias vsplit="tmux split-window -h"
+  alias hsplit="tmux split-window -v"
+  alias close="tmux kill-pane -t"
+
+fi
